chore: added performance overview task
